update cqlFilter for map widget that depends on table widget
