+++ web/client/components/widgets/enhancers/dependenciesToCqlFilter.js
@@ -1,0 +1,59 @@
+/*
+ * Copyright 2017, GeoSolutions Sas.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+const { compose, withPropsOnChange } = require('recompose');
+const { find, isEmpty} = require('lodash');
+
+const FilterUtils = require('../../../utils/FilterUtils');
+const { composeAttributeFilters } = require('../../../utils/FilterUtils');
+const { optionsToVendorParams } = require('../../../utils/VendorParamsUtils');
+const { arrayUpdate } = require('../../../utils/ImmutableUtils');
+
+const {composeFilterObject} = require('./utils');
+
+/**
+ * Merges dependencies quickFilters and filter into a cql filter for a layer
+ * @returns {object} the map with layers updated or not
+ */
+module.exports = compose(
+    withPropsOnChange(
+        ({mapSync, dependencies = {} } = {}, nextProps = {}, filter) =>
+            mapSync !== nextProps.mapSync
+            || dependencies !== nextProps.dependencies
+            || filter !== nextProps.filter,
+        ({ mapSync, dependencies = {}, filter: filterObj, map} = {}) => {
+
+            const targetLayerName = dependencies && dependencies.layer && dependencies.layer.name;
+            const layerInCommon = find(map.layers, {name: targetLayerName});
+            let filterObjCollection = {};
+
+            if (mapSync && layerInCommon) {
+                if (dependencies.quickFilters) {
+                    filterObjCollection = {...filterObjCollection, ...composeFilterObject(filterObj, dependencies.quickFilters, dependencies.options)};
+                }
+                if (dependencies.filter) {
+                    filterObjCollection = {...filterObjCollection, ...composeAttributeFilters([filterObjCollection, dependencies.filter])};
+                }
+
+                if (!isEmpty(filterObjCollection) && FilterUtils.toCQLFilter(filterObjCollection)) {
+                    const cqlFilter = FilterUtils.toCQLFilter(filterObjCollection);
+                    const layersUpdatedWithCql = arrayUpdate(false, {...layerInCommon, params: optionsToVendorParams({ ...layerInCommon, params: {CQL_FILTER: cqlFilter}})}, {name: targetLayerName}, map.layers);
+                    return {
+                        map: {
+                            ...map,
+                            layers: layersUpdatedWithCql
+                        }
+                    };
+                }
+            }
+            return {
+                map
+            };
+        }
+    )
+
+);
