+++ web/client/components/widgets/enhancers/utils.js
@@ -5,7 +5,10 @@
  * This source code is licensed under the BSD-style license found in the
  * LICENSE file in the root directory of this source tree.
  */
-const {get, find} = require('lodash');
+
+const {get, find, reduce, isEmpty} = require('lodash');
+const { gridUpdateToQueryUpdate } = require('../../../utils/FeatureGridUtils');
+const { composeAttributeFilters } = require('../../../utils/FilterUtils');
 /**
  * Utility function to get the original layer from "layers" dependency, then, get the "params" object.
  */
@@ -19,5 +22,23 @@ module.exports = {
                 }),
                 "params",
                 {}
-            )
+            ),
+    composeFilterObject: (filterObj, quickFilters, options) => {
+        const quickFiltersForVisibleProperties = quickFilters && options &&
+            Object.keys(quickFilters)
+                .filter(qf => find(options.propertyName, f => f === qf))
+                .reduce((p, c) => {
+                    return {...p, [c]: quickFilters[c]};
+                }, {});
+
+        // Building new filterObj in order to and the two filters: old filterObj and quickFilter (from Table Widget)
+        const columnsFilters = reduce(quickFiltersForVisibleProperties, (cFilters, value, attribute) => {
+            return gridUpdateToQueryUpdate({attribute, ...value}, cFilters);
+        }, {});
+        if (!isEmpty(filterObj) || !isEmpty(columnsFilters)) {
+            const composedFilterFields = composeAttributeFilters([filterObj, columnsFilters]);
+            return {...filterObj, ...composedFilterFields};
+        }
+        return {};
+    }
 };
