+++ web/client/components/widgets/widget/DefaultWidget.jsx
@@ -26,7 +26,7 @@ const wfsTable = require('../enhancers/wfsTable');
 const dependenciesToFilter = require('../enhancers/dependenciesToFilter');
 const dependenciesToOptions = require('../enhancers/dependenciesToOptions');
 const dependenciesToWidget = require('../enhancers/dependenciesToWidget');
-const dependenciesToLayers = require('../enhancers/dependenciesToLayers');
+const dependenciesToCqlFilter = require('../enhancers/dependenciesToCqlFilter');
 const dependenciesToMapProp = require('../enhancers/dependenciesToMapProp');
 //
 // connect widgets to dependencies, remote services and add base icons/tools
@@ -45,7 +45,7 @@ const TextWidget = compose(
 
 const MapWidget = compose(
     dependenciesToWidget,
-    dependenciesToLayers,
+    dependenciesToCqlFilter,
     dependenciesToMapProp('center'),
     dependenciesToMapProp('zoom'),
     mapWidget
